refactor batch apply engine layout
