+++ packages/slate/src/core/normalize-node.ts
@@ -1,6 +1,7 @@
 import { WithEditorFirstArg } from '../utils/types'
 import {
   Editor,
+  EditorNormalizeOptions,
   Element,
   Descendant,
   Text,
@@ -10,11 +11,57 @@ import {
   Ancestor,
 } from '../interfaces'
 
+const isDirectChildPath = (parentPath: Path, childPath: Path) =>
+  childPath.length === parentPath.length + 1 &&
+  Path.equals(Path.parent(childPath), parentPath)
+
+const getBlockOnlyChildIndexesToValidate = (
+  path: Path,
+  operation?: EditorNormalizeOptions['operation']
+) => {
+  if (!operation) {
+    return null
+  }
+
+  switch (operation.type) {
+    case 'set_node':
+    case 'insert_node': {
+      return isDirectChildPath(path, operation.path)
+        ? [operation.path[path.length]]
+        : null
+    }
+
+    case 'remove_node': {
+      return isDirectChildPath(path, operation.path) ? [] : null
+    }
+
+    case 'move_node': {
+      const removesFromParent = isDirectChildPath(path, operation.path)
+      const insertsIntoParent = isDirectChildPath(path, operation.newPath)
+
+      if (!removesFromParent && !insertsIntoParent) {
+        return null
+      }
+
+      if (removesFromParent && insertsIntoParent) {
+        return []
+      }
+
+      return insertsIntoParent ? [operation.newPath[path.length]] : []
+    }
+
+    default: {
+      return null
+    }
+  }
+}
+
 export const normalizeNode: WithEditorFirstArg<Editor['normalizeNode']> = (
   editor,
   entry,
   options
 ) => {
+  const { force = false, operation } = options ?? {}
   const [node, path] = entry as [{}, Path] // node is not yet normalized, treat as hostile
 
   // There are no core normalizations for text nodes.
@@ -113,21 +160,48 @@ export const normalizeNode: WithEditorFirstArg<Editor['normalizeNode']> = (
       }
     }
   } else {
+    const isInlineChild = (child: Descendant) =>
+      Element.isElement(child) && editor.isInline(child)
+
+    const normalizeBlockChild = (index: number) => {
+      const child = element.children[index]
+
+      if (!child || (!Node.isText(child) && !isInlineChild(child))) {
+        return
+      }
+
+      if (options?.fallbackElement) {
+        Transforms.wrapNodes(editor, options.fallbackElement(), {
+          at: path.concat(index),
+          voids: true,
+        })
+      } else {
+        Transforms.removeNodes(editor, { at: path.concat(index), voids: true })
+      }
+
+      element = Node.get(editor, path) as Ancestor
+    }
+
+    const childIndexesToValidate = force
+      ? null
+      : getBlockOnlyChildIndexesToValidate(path, operation)
+
+    if (Array.isArray(childIndexesToValidate)) {
+      for (const index of childIndexesToValidate) {
+        normalizeBlockChild(index)
+      }
+
+      return
+    }
+
     // Since we'll be applying operations while iterating, we also modify `n` when adding/removing nodes
     for (let n = 0; n < element.children.length; n++) {
-      const child = element.children[n]
-
       // Allow only block nodes in the top-level children and parent blocks that only contain block nodes
-      if (Node.isText(child) || editor.isInline(child)) {
-        if (options?.fallbackElement) {
-          Transforms.wrapNodes(editor, options.fallbackElement(), {
-            at: path.concat(n),
-            voids: true,
-          })
-        } else {
-          Transforms.removeNodes(editor, { at: path.concat(n), voids: true })
-        }
-        element = Node.get(editor, path) as Ancestor
+      if (
+        Node.isText(element.children[n]) ||
+        isInlineChild(element.children[n])
+      ) {
+        normalizeBlockChild(n)
         n--
       }
     }
